Refactor Client package structure
