+++ Snake_Game/src/client/Snake/Command/PlayerMoveCommand.java
@@ -1,6 +1,6 @@
-package client.Snake.Renderer.Command;
+package client.Snake.Command;
 
-import client.Snake.Renderer.Interface.ICommand;
+import client.Snake.Interface.ICommand;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
 import server.Snake.Enumerator.EPacketHeader;
@@ -9,72 +9,56 @@ import server.Snake.Network.Packet.Packet;
 import java.io.OutputStreamWriter;
 import java.util.HashMap;
 
-public class NetworkCommand extends TemplateCommand{
+public class PlayerMoveCommand extends TemplateCommand{
     private static ObjectWriter objectMapper = new ObjectMapper().writer();
     private static HashMap<Object, Object> packetMap = new HashMap<>();
     private static Packet packet = new Packet(EPacketHeader.CLIENT_RESPONSE);
     private static ICommand action;
 
-    public static void requestLogin(String id, OutputStreamWriter out){
-        action = new RequestLogin();
+    public static void moveUp(String id, OutputStreamWriter out){
+        action = new MoveUp();
         action.execute(id, out);
     }
 
-    public static void requestLogout(String id, OutputStreamWriter out){
-        action = new RequestLogout();
+    public static void moveDown(String id, OutputStreamWriter out){
+        action = new MoveDown();
         action.execute(id, out);
     }
 
-    public static void requestMatchJoin(String id, OutputStreamWriter out){
-        action = new RequestMatchJoin();
+    public static void moveRight(String id, OutputStreamWriter out){
+        action = new MoveRight();
         action.execute(id, out);
     }
 
-    public static void requestMatchLeave(String id, OutputStreamWriter out){
-        action = new RequestMatchLeave();
+    public static void moveLeft(String id, OutputStreamWriter out){
+        action = new MoveLeft();
+        action.execute(id, out);
+    }
+
+    public static void moveStop(String id, OutputStreamWriter out){
+        action = new MoveStop();
         action.execute(id, out);
     }
 
     @Override
     public void undo(String id, OutputStreamWriter out){
         action.undo(id, out);
+        //System.out.println(action.toString());
     }
     @Override
     public String getString(){
         return "current command: " + action.toString();
+
     }
 
-    private static class RequestLogin implements ICommand{
-        @Override
-        public void execute(String id, OutputStreamWriter out) {
-            packetMap.clear();
-
-            try {
-                packet = new Packet(EPacketHeader.CLIENT_LOGIN);
-                packet.setBody(objectMapper.writeValueAsString(packetMap));
-                out.write(packet.toString());
-                out.flush();
-            } catch (Exception e) {
-                System.out.println("Error sending an input to the server.");
-//                e.printStackTrace();
-            }
-        }
-
-        @Override
-        public void undo(String id, OutputStreamWriter out) {
-            action = new RequestLogout();
-            action.execute(id, out);
-        }
-    }
-
-    private static class RequestLogout implements ICommand{
+    private static class MoveUp implements ICommand{
         @Override
         public void execute(String id, OutputStreamWriter out) {
             packetMap.clear();
             packetMap.put("id", id);
-
+            packetMap.put("directionX", "0");
+            packetMap.put("directionY", "-1");
             try {
-                packet = new Packet(EPacketHeader.CLIENT_LOGOUT);
                 packet.setBody(objectMapper.writeValueAsString(packetMap));
                 out.write(packet.toString());
                 out.flush();
@@ -82,46 +66,50 @@ public class NetworkCommand extends TemplateCommand{
                 System.out.println("Error sending an input to the server.");
 //                e.printStackTrace();
             }
+
         }
 
         @Override
         public void undo(String id, OutputStreamWriter out) {
-            action = new RequestLogin();
+            action = new MoveLeft();
             action.execute(id, out);
         }
     }
 
-    private static class RequestMatchJoin implements ICommand{
-        @Override
-        public void execute(String id, OutputStreamWriter out) {
-            packetMap.clear();
-
-            try {
-                packet = new Packet(EPacketHeader.CLIENT_REQUEST_MATCH_JOIN);
-                packet.setBody(objectMapper.writeValueAsString(packetMap));
-                out.write(packet.toString());
-                out.flush();
-            } catch (Exception e) {
-                System.out.println("Error sending an input to the server.");
-//                e.printStackTrace();
-            }
-        }
-
-        @Override
-        public void undo(String id, OutputStreamWriter out) {
-            action = new RequestMatchLeave();
-            action.execute(id, out);
-        }
-    }
-
-    private static class RequestMatchLeave implements ICommand{
+    private static class MoveDown implements ICommand {
         @Override
         public void execute(String id, OutputStreamWriter out) {
             packetMap.clear();
             packetMap.put("id", id);
-
+            packetMap.put("directionX", "0");
+            packetMap.put("directionY", "1");
             try {
-                packet = new Packet(EPacketHeader.CLIENT_REQUEST_MATCH_LEAVE);
+                packet.setBody(objectMapper.writeValueAsString(packetMap));
+                out.write(packet.toString());
+                out.flush();
+            } catch (Exception e) {
+                System.out.println("Error sending an input to the server.");
+//                e.printStackTrace();
+            }
+
+        }
+
+        @Override
+        public void undo(String id, OutputStreamWriter out) {
+            action = new MoveRight();
+            action.execute(id, out);
+        }
+
+    }
+
+    private static class MoveRight implements ICommand{
+        @Override
+        public void execute(String id, OutputStreamWriter out) {
+            packetMap.clear();
+            packetMap.put("id", id);
+            packetMap.put("directionX", "1");
+            packetMap.put("directionY", "0");
+            try {
                 packet.setBody(objectMapper.writeValueAsString(packetMap));
                 out.write(packet.toString());
                 out.flush();
@@ -133,8 +121,60 @@ public class NetworkCommand extends TemplateCommand{
 
         @Override
         public void undo(String id, OutputStreamWriter out) {
-            action = new RequestMatchJoin();
+            action = new MoveUp();
             action.execute(id, out);
         }
+
     }
+
+    private static class MoveLeft implements ICommand{
+        @Override
+        public void execute(String id, OutputStreamWriter out) {
+            packetMap.clear();
+            packetMap.put("id", id);
+            packetMap.put("directionX", "-1");
+            packetMap.put("directionY", "0");
+            try {
+                packet.setBody(objectMapper.writeValueAsString(packetMap));
+                out.write(packet.toString());
+                out.flush();
+            } catch (Exception e) {
+                System.out.println("Error sending an input to the server.");
+//                e.printStackTrace();
+            }
+        }
+
+        @Override
+        public void undo(String id, OutputStreamWriter out) {
+            action = new MoveDown();
+            action.execute(id, out);
+        }
+
+    }
+
+    private static class MoveStop implements ICommand{
+        @Override
+        public void execute(String id, OutputStreamWriter out) {
+            packetMap.clear();
+            packetMap.put("id", id);
+            packetMap.put("directionX", "0");
+            packetMap.put("directionY", "0");
+            try {
+                packet.setBody(objectMapper.writeValueAsString(packetMap));
+                out.write(packet.toString());
+                out.flush();
+            } catch (Exception e) {
+                System.out.println("Error sending an input to the server.");
+//                e.printStackTrace();
+            }
+        }
+
+        @Override
+        public void undo(String id, OutputStreamWriter out) {
+            action = new MoveUp();
+            action.execute(id, out);
+        }
+
+    }
+
 }
