+++ Snake_Game/src/client/Snake/Interface/IDrawable.java
@@ -1,4 +1,4 @@
-package client.Snake.Renderer.Interface;
+package client.Snake.Interface;
 
 import java.awt.*;
