+++ Snake_Game/src/client/Snake/Interface/IMediator.java
@@ -1,4 +1,4 @@
-package client.Snake.Renderer.Interface;
+package client.Snake.Interface;
 
 public interface IMediator {
     void notify(Object sender);
