+++ Snake_Game/src/client/Snake/Renderer/Panels/SettingsPanel.java
@@ -1,24 +1,25 @@
-package client.Snake.Renderer;
+package client.Snake.Renderer.Panels;
 
 import client.Snake.Renderer.Components.MenuButton;
-import client.Snake.Renderer.Enumerator.ERendererState;
-import client.Snake.Renderer.Interface.IMediator;
+import client.Snake.Enumerator.ERendererState;
+import client.Snake.Interface.IMediator;
+import client.Snake.Renderer.SwingRender;
 
 import javax.swing.*;
 
-public class HostGamePanel extends JPanel implements Runnable, IMediator {
+public class SettingsPanel extends JPanel implements Runnable, IMediator {
     private static SwingRender render = null;
-    private static HostGamePanel panelInstance = null;
+    private static SettingsPanel panelInstance = null;
 
     private static MenuButton backButton;
 
-    private HostGamePanel() {
+    private SettingsPanel() {
         setUpControls();
     }
 
-    public static HostGamePanel getInstance() {
+    public static SettingsPanel getInstance() {
         if (panelInstance == null)
-            panelInstance = new HostGamePanel();
+            panelInstance = new SettingsPanel();
         return panelInstance;
     }
 
@@ -33,13 +34,13 @@ public class HostGamePanel extends JPanel implements Runnable, IMediator {
 
     @Override
     public void notify(Object sender) {
-        if (this.backButton.equals(sender)) {
-            this.render.setCurrentState(ERendererState.MENU);
-            this.render.remove(this);
+        if (backButton.equals(sender)) {
+            render.setCurrentState(ERendererState.MENU);
+            render.remove(this);
             this.invalidate();
             this.validate();
-            synchronized (this.render) {
-                this.render.notify();
+            synchronized (render) {
+                render.notify();
             }
         }
     }
